requirements.txt
# Backend dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
alembic>=1.13.0
pydantic[email]>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.1.0
pyjwt>=2.8.0
bcrypt>=4.1.0
python-multipart>=0.0.6
httpx>=0.26.0
pytest>=7.4.0
pytest-asyncio>=0.23.0
